Fix bug with nullable rules (found by github.com/ndw)
The original code that I wrote had some issues with the nullable grammar
rules.

Constructing the set of nullable rules in the grammar had a bug where
the rule would be marked as nullable if any of the symbols were
nullable. The proper behavior is to have it marked as nullable if all
the symbols are nullable.

The complete/predict step after walking a nullable symbol also had a bug
- it would check all symbols to see if they were nullable and add them
to the next state set. The correct behavior is only add nullable symbols
that follow the symbol until there's a non-nullable symbol.